Sorting of Files, addition of a moving ROI for the Fall
Changes were made to the region of interest for the fall. It is now possible to automatically change the ROI by shifting it by the displacemnt from the previous image pair. This part is still under development but should work fine.

Also the files were sorted for Federica and for developing.