Finalizing ROI prediction
Added the new prediction of the ROI using a padded velocity field that has the boundary conditions. The calculation is done using trapz to get the mean velocity. To eliminate outliers the bottom 5 rows are used in the calculation.

Also added a maximum sig2noise ratio that is tolerated, this can not be set in the client, has to be changed in validation_patch. Also added validation_patch, that accounts for these new settings